requirements.txt
pyparsing
gmpy2
numpy
pydot
psutil
pycosat
antlr4-python3-runtime
